+++ hooks/post_gen_project.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python
 import os
 
 PROJECT_DIRECTORY = os.path.realpath(os.path.curdir)
@@ -20,6 +19,8 @@ if __name__ == '__main__':
     if 'no' in '{{ cookiecutter.command_line_interface|lower }}':
         cli_file = os.path.join('{{ cookiecutter.project_slug }}', 'cli.py')
         remove_file(cli_file)
+        cli_bash_autocomplete_file = os.path.join('{{ cookiecutter.project_slug }}', 'dev-util', 'bash_autocomplete.sh')
+        remove_file(cli_bash_autocomplete_file)
 
     if 'Not open source' == '{{ cookiecutter.open_source_license }}':
         remove_file('LICENSE')