+++ tests/test_bake_project.py
@@ -115,7 +115,7 @@ def test_bake_with_apostrophe_and_run_tests(cookies):
 #
 #         # when:
 #         travis_setup_cmd = ('python travis_pypi_setup.py'
-#                             ' --repo audreyr/cookiecutter-pypackage --password invalidpass')
+#                             ' --repo thomasjahoda/cookiecutter-pypackage --password invalidpass')
 #         run_inside_dir(travis_setup_cmd, project_path)
 #         # then:
 #         result_travis_config = yaml.load(result.project.join(".travis.yml").open())
@@ -205,7 +205,7 @@ def test_not_using_pytest(cookies):
 #
 #     # when:
 #     travis_setup_cmd = ('python travis_pypi_setup.py'
-#                         ' --repo audreyr/cookiecutter-pypackage --password invalidpass')
+#                         ' --repo thomasjahoda/cookiecutter-pypackage --password invalidpass')
 #     run_inside_dir(travis_setup_cmd, project_path)
 #
 #     # then: