+++ {{cookiecutter.project_slug}}/tests/test_{{cookiecutter.project_slug}}.py
@@ -1,8 +1,3 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-
-"""Tests for `{{ cookiecutter.project_slug }}` package."""
-
 {% if cookiecutter.use_pytest == 'y' -%}
 import pytest
 {% else %}
@@ -51,16 +46,7 @@ def test_command_line_interface():
 
 
 class Test{{ cookiecutter.project_slug|title }}(unittest.TestCase):
-    """Tests for `{{ cookiecutter.project_slug }}` package."""
 
-    def setUp(self):
-        """Set up test fixtures, if any."""
-
-    def tearDown(self):
-        """Tear down test fixtures, if any."""
-
-    def test_000_something(self):
-        """Test something."""
 {%- if cookiecutter.command_line_interface|lower == 'click' %}
 
     def test_command_line_interface(self):